Refactor: relocate exceptions to services package
