+++ desafio06-RestAssured/src/main/java/com/devsuperior/dsmovie/services/exceptions/ResourceNotFoundException.java
